Refactor ppl creator to decouple PipelineGenerator from PipelineConfigGenerator
