+++ tools/dynamic_ppl_runner/.gitignore
@@ -1,0 +1,3 @@
+dlsps-config.json
+.env
+__pycache__/
